Using generalized code runner
