+++ src/Server/Test.hs
@@ -9,7 +9,6 @@ import           Server.Test.ExpectationsRunner
 import           Server.Test.SmellsDetectorRunner
 import           Server.Test.RequestValidator
 
-
 process :: Request -> IO Response
 process r
   | (Just message) <- validateRequest r = return emptyResponse { exit = "aborted", out =  message }
@@ -24,9 +23,9 @@ run r = do
         smellsResuls = runSmellsDetection (content r)
         totalExpectationResults = expectationResults ++ smellsResuls
 
-toResponse :: Either TestError TestResults -> Response
-toResponse (Left (e, o)) = emptyResponse { exit = e, out = o }
-toResponse (Right trs)   = emptyResponse { testResults = trs }
+toResponse :: RunnerResult TestResults -> Response
+toResponse (Error (e, o)) = emptyResponse { exit = e, out = o }
+toResponse (Ok trs)   = emptyResponse { testResults = trs }
 
 compileRequest :: Request -> String
 compileRequest request = compile (test request) (extra request) (content request)
